requirement.txt
streamlit
pandas
numpy
scikit-learn
xgboost
torch
transformers
nltk
Pillow
timm